pequeños arreglos..

+++ Practicas/Practica2/freeglut project 2D/freeglut project 2D/src/Obstacle.cpp
@@ -10,6 +10,6 @@ Obstacle::~Obstacle(void)
 {
 }
 
-Obstacle::Obstacle(PV2D *c,int n, GLdouble radius, GLdouble o,GLdouble r, GLdouble g, GLdouble b):
+Obstacle::Obstacle(PV2D *c,int n, GLdouble radius, GLdouble o,GLfloat r, GLfloat g, GLfloat b):
 	DrawablePolygon(c,n,radius,o,r,g,b){
 }
